+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- fix(usePreconditionFunc): 修改错误拼写，修改类型，支持getter函数
+
 ## 1.4.0(2024-8-4)
 
 - usePreconditionFunc(feat): 加入副作用函数调用hook